+++ core/data-model/src/main/java/com/ivy/core/data/Budget.kt
@@ -1,15 +1,19 @@
 package com.ivy.core.data
 
-import com.ivy.core.data.attribute.CategoryId
-import java.time.LocalDateTime
+import com.ivy.core.data.common.ItemIconId
+import com.ivy.core.data.common.IvyColor
+import com.ivy.core.data.common.Reorderable
+import com.ivy.core.data.common.Value
 import java.util.*
 
 data class Budget(
     val id: UUID,
-    val visuals: ItemVisuals,
+    val name: String,
+    val description: String?,
+    val iconId: ItemIconId,
+    val color: IvyColor,
     val amount: Value,
     val carryOver: Boolean,
-    val startDate: LocalDateTime,
     val interval: BudgetInterval,
     val categories: List<CategoryId>,
     override val orderNum: Double
