+++ core/data-model/src/main/java/com/ivy/core/data/Category.kt
@@ -1,13 +1,17 @@
-package com.ivy.core.data.attribute
+package com.ivy.core.data
 
-import com.ivy.core.data.Archiveable
-import com.ivy.core.data.ItemVisuals
-import com.ivy.core.data.Reorderable
+import com.ivy.core.data.common.Archiveable
+import com.ivy.core.data.common.ItemIconId
+import com.ivy.core.data.common.IvyColor
+import com.ivy.core.data.common.Reorderable
 import java.util.*
 
 data class Category(
     val id: UUID,
-    val visuals: ItemVisuals,
+    val name: String,
+    val description: String?,
+    val iconId: ItemIconId,
+    val color: IvyColor,
     val type: CategoryType,
     val parentCategory: CategoryId?,
     override val orderNum: Double,
