+++ core/data-model/src/main/java/com/ivy/core/data/Folder.kt
@@ -1,4 +1,4 @@
-package com.ivy.core.data.attribute
+package com.ivy.core.data
 
 import java.util.*
