+++ core/data-model/src/main/java/com/ivy/core/data/RecurringRule.kt
@@ -1,0 +1,3 @@
+package com.ivy.core.data
+
+// TODO: Model
