--- core/data-model/src/main/java/com/ivy/core/data/attribute/Tag.kt
@@ -1,16 +1,0 @@
-package com.ivy.core.data.attribute
-
-import com.ivy.core.data.Archiveable
-import com.ivy.core.data.ItemVisuals
-import com.ivy.core.data.Reorderable
-import java.util.*
-
-data class Tag(
-    val id: UUID,
-    val visuals: ItemVisuals,
-    override val orderNum: Double,
-    override val archived: Boolean,
-) : Reorderable, Archiveable
-
-@JvmInline
-value class TagId(val id: UUID)
