+++ core/data-model/src/main/java/com/ivy/core/data/common/Archiveable.kt
@@ -1,4 +1,4 @@
-package com.ivy.core.data
+package com.ivy.core.data.common
 
 interface Archiveable {
     val archived: Boolean
